Refactor message-merging logic into separate utility.
